GH-52: Lambda fetch req error handling
