+++ src/lambda/src/config/config.js
@@ -1,13 +1,13 @@
 const clientAppUrl = 'https://www.stations.ie';
-const localhostRegEx = /https?:\/\/localhost:\d{4}/;
-const netlifyDeployPreviewRegEx = /https:\/\/deploy-preview-\d{1,3}--stations.netlify.com/;
-const netlifyDeployPreviewShaRegEx = /https:\/\/[a-z0-9]{24}--stations.netlify.com/;
+const localhostUrls = /https?:\/\/localhost:\d{4}/;
+const netlifyDeployPreviewUrls = /https:\/\/deploy-preview-\d{1,3}--stations.netlify.com/;
+const netlifyDeployPreviewShaUrls = /https:\/\/[a-z0-9]{24}--stations.netlify.com/;
 
 const corsWhiteList = [
   clientAppUrl,
-  localhostRegEx,
-  netlifyDeployPreviewRegEx,
-  netlifyDeployPreviewShaRegEx
+  localhostUrls,
+  netlifyDeployPreviewUrls,
+  netlifyDeployPreviewShaUrls
 ];
 
 module.exports = {
