+++ src/lambda/src/utility/utils.js
@@ -1,13 +1,19 @@
+'use strict';
+
 const xml2js = require('xml2js');
 
-const parser = new xml2js.Parser({explicitArray: false});
+const xml2jsParserOptions = {
+  explicitArray: false
+};
+
+const parser = new xml2js.Parser(xml2jsParserOptions);
 
 const parseXmlBody = (xml) => {
   return new Promise((resolve, reject) => {
     parser.parseString(xml, (error, result) => {
       if (error) {
         reject({
-          parseError: {
+          xmlParseError: {
             error: error,
             message: 'xml parse error'
           }
@@ -41,17 +47,31 @@ const filterByKey = (obj, filter) => {
 };
 
 // https://medium.com/front-end-weekly/30-seconds-of-code-rename-many-object-keys-in-javascript-268f279c7bfa
-const renameKeys = (obj, mapKeys) => {
+const renameKeys = (obj, keysMapping) => {
   return Object.keys(obj)
     .reduce((acc, key) => ({
       ...acc,
-      ...{[mapKeys[key] || key]: obj[key]}
+      ...{[keysMapping[key] || key]: obj[key]}
     }), {});
+};
+
+const checkResponseStatus = (res) => {
+  if (res.ok) {
+    return res;
+  } else {
+    return Promise.reject({
+      statusError: {
+        error: res.status,
+        message: res.statusText
+      }
+    });
+  }
 };
 
 module.exports = {
   getRequiredProps,
   filterByKey,
   renameKeys,
-  parseXmlBody
+  parseXmlBody,
+  checkResponseStatus
 };
